lib/azure_f/requirements.txt
azure-functions
websockets
brotli

# Test:
# pytest
